update signature
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3812 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/HelpViewer/BRCell.m
@@ -47,7 +47,7 @@ static BRCell *brCell;
 -(NSRect) cellFrameForTextContainer: (NSTextContainer *)c
     proposedLineFragment: (NSRect)lf
     glyphPosition: (NSPoint)p
-    characterIndex: (unsigned int)ci
+    characterIndex: (NSUInteger)ci
 {
     NSNumber *width,*height;
     CGFloat w,h;
